+++ src/functions/sort-array-by-object-property.ts
@@ -1,14 +1,22 @@
 /**
  * Sort an array of objects by one of the object's properties (in-place).
  * @param array - The array to sort.
- * @param property - The property name to sort by. (can be property.property)
+ * @param attributeGetter - The property name to sort by. (can be property.property)
  * @param inverse - Inverse the output (descending).
  */
-export const sortArrayByObjectProperty = <T>(array: T[], property: string, inverse = false): T[] => {
+// eslint-disable-next-line max-len,sonarjs/cognitive-complexity
+export const sortArrayByObjectProperty = <T>(array: T[], attributeGetter: string | ((item: T) => string | number | undefined), inverse = false): T[] => {
+
+    let propertyIsString = false;
+    if (typeof attributeGetter === 'string') {
+        // eslint-disable-next-line no-console,max-len
+        console.warn('Using sortArrayByObjectProperty(T[], string, boolean) is deprecated, use a selector function as the second parameter.');
+        propertyIsString = true;
+    }
 
     const getPropertyFromPath = (value: any): any => {
         const getObjectValue = (object: any, key: string): any => object[key];
-        return property.split('.').reduce(getObjectValue, value);
+        return (attributeGetter as string).split('.').reduce(getObjectValue, value);
     };
 
     const checkIfEqualTypes = (left: any, right: any): void => {
@@ -22,8 +30,10 @@ export const sortArrayByObjectProperty = <T>(array: T[], property: string, inver
     };
 
     const compare = (a: any, b: any) => {
-        let left = getPropertyFromPath(a);
-        let right = getPropertyFromPath(b);
+        // @ts-ignore
+        let left = propertyIsString ? getPropertyFromPath(a) : attributeGetter(a);
+        // @ts-ignore
+        let right = propertyIsString ? getPropertyFromPath(b) : attributeGetter(b);
 
         checkIfEqualTypes(left, right);