+++ CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
+### Changed
 - objectsArrayToObject now takes a selector function, using a string as key is deprecated.
+- sortArrayByObjectProperty now takes a selector function, using a string as key is deprecated.
 
 ## [3.1.0] - 2020-05-18
 ### Added